Move cache profiler to dedicated namespace
